fix(ci): python-coverage-shards maintainability — split loop.py NotifyHuman arm under 1500-line limit
loop.py exceeded the 1500-line first-party file guardrail (1531 lines),
failing tests/unit/test_core_decomposition_maintainability.py. Extract the
NotifyHuman dispatch arm into a sibling notify_human_loop module. The loop
keeps the isinstance(action, NotifyHuman) dispatch check and returns the
helper's result, so the helper takes a concrete NotifyHuman and the coverage
profile is unchanged (the check's false branch still falls through to the
existing pragma-no-cover guard). Behavior unchanged; loop.py is now 1279
lines and the helper 314.
